move form-related stuff in events into separate file
